Refactor Interpretation edit/dialog UI
*put interpretation edit component in interpretation dialog
*remove interpretation edit
*move all interpretation sub components inside of interpretation dialog component
