+++ phenopacketlab-frontend/src/app/component/pheno-creator/interpretation-step.component.html
@@ -50,7 +50,7 @@
               pRipple
               icon="pi pi-pencil"
               class="p-button-rounded p-button-text"
-              (click)="addInterpretation(interpretation)"
+              (click)="editInterpretation(interpretation)"
             ></button>
             <button
               pButton
