+++ phenopacketlab-frontend/src/app/component/pheno-creator/interpretation-step.component.ts
@@ -10,7 +10,7 @@ import { Profile, ProfileSelection } from 'src/app/models/profile';
 import { InterpretationService } from 'src/app/services/interpretation.service';
 import { PhenopacketService } from 'src/app/services/phenopacket.service';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { InterpretationDialogComponent } from '../shared/dialog/interpretation-dialog/interpretation-dialog.component';
+import { InterpretationDialogComponent, InterpretationDialogMode } from '../shared/dialog/interpretation-dialog/interpretation-dialog.component';
 import { Utils } from '../shared/utils';
 
 @Component({
@@ -82,12 +82,10 @@ export class InterpretationStepComponent implements OnInit, OnDestroy {
         }
     }
 
-    addInterpretation(interpretation?: Interpretation) {
-        if (interpretation === undefined || interpretation === null) {
-            interpretation = new Interpretation();
-        }
+    addInterpretation() {
+        const interpretation = new Interpretation();
         this.ref = this.dialogService.open(InterpretationDialogComponent, {
-            header: 'Enter Interpretation',
+            header: 'Add Interpretation',
             width: '70%',
             contentStyle: { 'overflow': 'auto' },
             baseZIndex: 10000,
@@ -97,28 +95,54 @@ export class InterpretationStepComponent implements OnInit, OnDestroy {
             data: {
                 interpretation: interpretation,
                 phenopacket: this.phenopacket,
-                profile: this.profileSelection
+                profile: this.profileSelection,
+                mode: InterpretationDialogMode.ADD
             }
         });
-
         this.ref.onClose.subscribe((interpret: Interpretation) => {
-            if (interpret) {
-                const indexToUpdate = this.interpretations.findIndex(item => item.id === interpret.id);
-                interpret.key = Utils.getBiggestKey(this.interpretations) + 1;
-                if (indexToUpdate === -1) {
-                    this.interpretations.push(interpret);
-                } else {
-                    this.interpretations[indexToUpdate] = interpret;
-                    this.interpretations = Object.assign([], this.interpretations);
-                }
-                // this.showTable = true;
-                // emit change
-                this.phenopacket.interpretations = this.interpretations;
-                this.submitted = true;
-            }
+            this.updateInterpretation(interpret);
         });
     }
 
+    editInterpretation(interpretation: Interpretation) {
+        this.ref = this.dialogService.open(InterpretationDialogComponent, {
+            header: 'Edit Interpretation',
+            width: '70%',
+            contentStyle: { 'overflow': 'auto' },
+            baseZIndex: 10000,
+            resizable: true,
+            draggable: true,
+            modal: true,
+            data: {
+                interpretation: interpretation,
+                phenopacket: this.phenopacket,
+                profile: this.profileSelection,
+                mode: InterpretationDialogMode.EDIT
+            }
+        });
+        this.ref.onClose.subscribe((interpret: Interpretation) => {
+            this.updateInterpretation(interpret);
+        });
+    }
+
+    updateInterpretation(interpretation: Interpretation) {
+        if (interpretation) {
+            const indexToUpdate = this.interpretations.findIndex(item => item.id === interpretation.id);
+            interpretation.key = Utils.getBiggestKey(this.interpretations) + 1;
+            if (indexToUpdate === -1) {
+                this.interpretations.push(interpretation);
+            } else {
+                this.interpretations[indexToUpdate] = interpretation;
+                this.interpretations = Object.assign([], this.interpretations);
+            }
+            console.log(interpretation);
+            console.log(this.interpretations);
+            // this.showTable = true;
+            // emit change
+            this.phenopacket.interpretations = this.interpretations;
+            this.submitted = true;
+        }
+    }
     deleteInterpretation(interpretation: Interpretation) {
         this.confirmationService.confirm({
             message: 'Are you sure you want to delete \'' + interpretation.id + '\'?',
@@ -139,12 +163,12 @@ export class InterpretationStepComponent implements OnInit, OnDestroy {
         });
     }
 
-    updateInterpretation(interpretation) {
-        if (this.interpretations === undefined) {
-            this.interpretations = [];
-        }
-        this.interpretations.push(interpretation);
-    }
+    // updateInterpretation(interpretation) {
+    //     if (this.interpretations === undefined) {
+    //         this.interpretations = [];
+    //     }
+    //     this.interpretations.push(interpretation);
+    // }
 
     /**
      * Called when a row is selected in the left side table
