+++ phenopacketlab-frontend/src/app/component/phenopacket/interpretation/interpretation.component.html
@@ -5,6 +5,19 @@
       dataKey="key"
       styleClass="phenopacket-list-tabs"
     >
+      <ng-template pTemplate="caption">
+        <div class="flex align-items-center justify-content-between">
+          Added interpretation(s)
+          <button
+            pButton
+            icon="pi pi-plus"
+            label="Add interpretation"
+            class="p-button-text"
+            style="width: 250px"
+            (click)="addInterpretation()"
+          ></button>
+        </div>
+      </ng-template>
       <ng-template pTemplate="header">
         <tr>
           <th style="width: 5rem"></th>
@@ -37,7 +50,7 @@
               pRipple
               icon="pi pi-pencil"
               class="p-button-rounded p-button-text"
-              (click)="addInterpretation(interpretation)"
+              (click)="editInterpretation(interpretation)"
             ></button>
             <button
               pButton
@@ -62,20 +75,12 @@
       </ng-template>
       <ng-template pTemplate="emptymessage">
         <tr>
-          <td [attr.colspan]="6">
-            No interpretation to display.
-          </td>
-        <tr>
-     </ng-template>
+          <td [attr.colspan]="6">No interpretation to display.</td>
+        </tr>
+
+        <tr></tr
+      ></ng-template>
     </p-table>
   </div>
 </div>
-<button
-  pButton
-  label="Add interpretation"
-  (click)="addInterpretation()"
-  icon="pi pi-plus"
-  class="p-button-text"
-  style="width: 250px"
-></button>
 <p-confirmDialog header="Confirmation" width="425"></p-confirmDialog>
