+++ phenopacketlab-frontend/src/app/component/phenopacket/interpretation/interpretation.component.ts
@@ -5,7 +5,7 @@ import { Interpretation } from 'src/app/models/interpretation';
 import { ConfirmationService, MessageService } from 'primeng/api';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { Phenopacket } from 'src/app/models/phenopacket';
-import { InterpretationDialogComponent } from '../../shared/dialog/interpretation-dialog/interpretation-dialog.component';
+import { InterpretationDialogComponent, InterpretationDialogMode } from '../../shared/dialog/interpretation-dialog/interpretation-dialog.component';
 
 @Component({
     selector: 'app-interpretation',
@@ -40,39 +40,55 @@ export class InterpretationComponent implements OnInit {
     ngOnInit() {
     }
 
-    /**
-     * Add a new interpretation(genomic) with default values or no values
-     */
-    addInterpretation(interpretation?: Interpretation) {
-        if (interpretation === undefined || interpretation === null) {
-            interpretation = new Interpretation();
-        }
+    addInterpretation() {
+        const interpretation = new Interpretation();
         this.ref = this.dialogService.open(InterpretationDialogComponent, {
-            header: 'Enter Interpretation',
-            width: '70%',
+            header: 'Add Interpretation',
+            width: '60%',
             contentStyle: { 'overflow': 'auto' },
             baseZIndex: 10000,
             resizable: true,
             draggable: true,
             data: { interpretation: interpretation,
-                    phenopacket: this.phenopacket }
+                    phenopacket: this.phenopacket,
+                    mode: InterpretationDialogMode.ADD }
         });
-
         this.ref.onClose.subscribe((interpret: Interpretation) => {
-            if (interpret) {
-                const indexToUpdate = this.interpretations.findIndex(item => item.id === interpret.id);
-                if (indexToUpdate === -1) {
-                    this.interpretations.push(interpret);
-                } else {
-                    this.interpretations[indexToUpdate] = interpret;
-                    this.interpretations = Object.assign([], this.interpretations);
-                }
-                // emit change
-                this.onInterpretationsChange.emit(this.interpretations);
-            }
+            this.updateInterpretation(interpret);
         });
     }
 
+    editInterpretation(interpretation: Interpretation) {
+
+        this.ref = this.dialogService.open(InterpretationDialogComponent, {
+            header: 'Edit Interpretation',
+            width: '60%',
+            contentStyle: { 'overflow': 'auto' },
+            baseZIndex: 10000,
+            resizable: true,
+            draggable: true,
+            data: { interpretation: interpretation,
+                    phenopacket: this.phenopacket,
+                    mode: InterpretationDialogMode.EDIT }
+        });
+        this.ref.onClose.subscribe((interpret: Interpretation) => {
+            this.updateInterpretation(interpret);
+        });
+    }
+
+    updateInterpretation(interpretation: Interpretation) {
+        if (interpretation) {
+            const indexToUpdate = this.interpretations.findIndex(item => item.id === interpretation.id);
+            if (indexToUpdate === -1) {
+                this.interpretations.push(interpretation);
+            } else {
+                this.interpretations[indexToUpdate] = interpretation;
+                this.interpretations = Object.assign([], this.interpretations);
+            }
+            // emit change
+            this.onInterpretationsChange.emit(this.interpretations);
+        }
+    }
     deleteInterpretation(interpretation: Interpretation) {
         this.confirmationService.confirm({
             message: 'Are you sure you want to delete \'' + interpretation.id + '\'?',
