+++ phenopacketlab-frontend/src/app/component/shared/dialog/interpretation-dialog/genomic-interpretation/gene-descriptor/gene-descriptor.component.ts
