+++ phenopacketlab-frontend/src/app/component/shared/dialog/interpretation-dialog/genomic-interpretation/gene-descriptor/label-creator-dialog.component.html
