+++ phenopacketlab-frontend/src/app/component/shared/dialog/interpretation-dialog/genomic-interpretation/variant-interpretation/variant-interpretation.component.ts
