+++ phenopacketlab-frontend/src/app/component/shared/dialog/interpretation-dialog/genomic-interpretation/variant-interpretation/variation-descriptor/variation-descriptor.component.scss
