--- phenopacketlab-frontend/src/app/component/shared/edit/interpretation-edit/interpretation-edit.component.scss
@@ -1,22 +1,0 @@
-.card {
-    overflow-x: auto;
-    overflow-y: auto;
-}
-
-.label-active:checked {
-    color: black;
-}
-
-.label-active {
-    color: red;
-}
-
-ul {
-    display: block;
-    list-style-type: none;
-    margin-top: 0;
-    margin-bottom: 0;
-    margin-left: 0;
-    margin-right: 0;
-    padding-left: 1rem;
-}
