--- phenopacketlab-frontend/src/app/component/shared/edit/interpretation-edit/interpretation-edit.component.spec.ts
@@ -1,47 +1,0 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { ButtonModule } from 'primeng/button';
-import { DropdownModule } from 'primeng/dropdown';
-import { DialogService, DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { RadioButtonModule } from 'primeng/radiobutton';
-import { SharedModule } from 'src/app/component/shared/shared.module';
-import { InterpretationEditComponent } from './interpretation-edit.component';
-
-describe('InterpretationEditComponent', () => {
-  let component: InterpretationEditComponent;
-  let fixture: ComponentFixture<InterpretationEditComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ InterpretationEditComponent ],
-      imports: [
-        ReactiveFormsModule,
-        NoopAnimationsModule,
-        FormsModule,
-        DropdownModule,
-        RadioButtonModule,
-        ButtonModule,
-        SharedModule,
-        HttpClientTestingModule
-      ],
-      providers: [
-        DynamicDialogRef,
-        DynamicDialogConfig,
-        DialogService
-      ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(InterpretationEditComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
